Add comments in code for doxygen usage.

+++ A-Eye_Controller/A-Eye_Controller/Program.cs
@@ -9,13 +9,33 @@ namespace AEye
     /// </summary>
     public static class Program
     {
+        /// <summary>
+        /// Different communication model can be used. 
+        /// Here you can see which one. 
+        /// </summary>
         public enum ComMode
         {
+            /// <summary>
+            /// Historical way of communication, you will use named pipe for TM, 
+            /// comparison between json file for TC, and TCP connection implemented in python. 
+            /// This is deprecated.
+            /// </summary>
             JSONxNAMEDPIPE_e,
+            /// <summary>
+            /// New way of communication, every data transits by MQTT. 
+            /// There's no more TCP client and python client are not used. 
+            /// Every communication are directly in C#.
+            /// Python is still used for naming and ordering images. 
+            /// </summary>
             MQTT_e
         }
 
-        public static ComMode comMode = ComMode.MQTT_e; // 0 for JSON, 1 for MQTT
+        /// <summary>
+        /// By setting this variable, you choose which communication you want to use. 
+        /// See enum type. 
+        /// </summary>
+        public static ComMode comMode = ComMode.MQTT_e;
+
         /// <summary>
         /// Store the IP address.
         /// </summary>
@@ -30,6 +50,8 @@ namespace AEye
         /// Store the log.
         /// </summary>
         public static string log = "";
+
+
         /// <summary>
         ///  The main entry point for the application.
         /// </summary>
@@ -39,28 +61,37 @@ namespace AEye
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
 
-            //DEBUG
+            /*// DEBUG --> This will kill the previous process.
             foreach (var process in Process.GetProcessesByName("pythonw"))
             {
                 process.Kill();
-            }
+            }*/
 
             ApplicationConfiguration.Initialize();
-            resetConfig(); 
+            if (comMode == ComMode.JSONxNAMEDPIPE_e)
+            {
+                resetConfig(); 
+            }
             Thread run_thr = new Thread(RunThread);
             run_thr.Start();
             Application.Run(controller);
 
-            // Clean spaces
-            Thread.Sleep(3000);
-            foreach (var process in Process.GetProcessesByName("pythonw"))
+            if (comMode == ComMode.JSONxNAMEDPIPE_e)
             {
-                process.Kill();
+                // Clean spaces
+                Thread.Sleep(3000);
+                foreach (var process in Process.GetProcessesByName("pythonw"))
+                {
+                    process.Kill();
+                }
             }
 
             Environment.Exit(0);
         }
 
+        /// <summary>
+        /// Initialise last_config.json with init parameters.
+        /// </summary>
         private static void resetConfig()
         {
             // Create empty configuration
